refactor(latest event): move the back-pagination to `RoomLatestEvents`
