+++ crates/matrix-sdk/src/event_cache/mod.rs
@@ -68,9 +68,7 @@ mod tasks;
 #[cfg(feature = "e2e-encryption")]
 pub use redecryptor::{DecryptionRetryRequest, RedecryptorReport};
 
-pub(crate) use self::back_pagination_queue::{
-    BATCH_SIZE, BackPaginationRequest, BackPaginationStopReason, Priority,
-};
+pub(crate) use self::back_pagination_queue::{BATCH_SIZE, BackPaginationRequest, Priority};
 pub use self::{
     back_pagination_queue::BackPaginationQueue,
     caches::{
